requirements.txt
streamlit

# LangChain split packages – we DO NOT import from `langchain.chains`
langchain-core
langchain-community
langchain-ibm

chromadb
pypdf
python-dotenv
ibm-watsonx-ai

